Settings/Preferences will no longer break when Update is pressed
