+++ .elasticbeanstalk/config.yml
@@ -4,6 +4,8 @@ branch-defaults:
     group_suffix: null
   user-preferences:
     environment: thymeleaf-starter-dev
+  user-prefs:
+    environment: thymeleaf-starter-dev
 deploy:
   artifact: build/libs/app.jar
 global:
