+++ src/main/resources/db/migration/V3.0__users.sql
@@ -1,0 +1,2 @@
+alter table `users`
+add column preferences varchar(200) not null;
